# packages/nocodb/src/ee/schema/swagger.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "nocodb",
    "version": "1.0",
    "description": "NocoDB API Documentation"
  },
...
  "servers": [
    {
      "url": "https://app.nocodb.com"
    }
  ],
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/users": {
      "get": {
        "summary": "Workspace users list",
        "operationId": "workspace-user-list",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceUserList"
                }
              }
            }
          }
        },
        "tags": [
          "Workspace user"
        ],
        "description": "Workspace users list"
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "workspaceId",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "boolean"
          },
          "name": "includeDeleted",
          "in": "query"
        }
      ]
    },
...
    "/api/v1/workspaces/{workspaceId}/invitations": {
      "post": {
        "summary": "Workspace user invite",
        "operationId": "workspace-user-invite",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUserInvite"
              }
            }
          }
        },
        "tags": [
          "Workspace user"
        ],
        "description": "Workspace user invite"
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "workspaceId",
          "in": "path",
          "required": true
        }
      ]
    },
...
    "/api/v1/workspaces/{workspaceId}/users/{userId}": {
      "get": {
        "summary": "Workspace user read",
        "operationId": "workspace-user-read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceUser"
                }
              }
            }
          }
        },
        "tags": [
          "Workspace user"
        ],
        "description": "Workspace user read"
      },
      "patch": {
        "summary": "Update workspace user",
        "operationId": "workspace-user-update",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Workspace user"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roles": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "description": "Update workspace user"
      },
      "delete": {
        "summary": "Delete workspace user",
        "operationId": "workspace-user-delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Workspace User"
        ],
        "description": "Delete workspace user"
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "workspaceId",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "userId",
          "in": "path",
          "required": true
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "WorkspaceUser": {
        "description": "",
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "fk_user_id": {
            "type": "string"
          },
          "invite_accepted": {
            "type": "boolean"
          },
          "invite_token": {
            "type": "string"
          },
          "roles": {
            "type": "string"
          }
        },
        "title": "Workspace User",
        "type": "object",
        "x-internal": false
      },
      "WorkspaceUserInvite": {
        "description": "",
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "roles": {
            "type": "string"
          }
        },
        "title": "Workspace User Invite",
        "type": "object",
        "x-internal": false
      },
      "WorkspaceUserList": {
        "description": "",
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceUser"
            },
            "type": "array"
          },
          "pageInfo": {
            "$ref": "#/components/schemas/Paginated"
          }
        },
        "type": "object"
      },
...
    }
  }
}